+++ js/jison/makeWrite.js
@@ -802,6 +802,7 @@
 				return indent + this.body.trim();
 			}
 		} else {
+			console.log("asda:"+this.body)
 			return indent + this.body;
 		}
 	};