+++ js/translate.js
@@ -135,12 +135,13 @@
 		document.getElementById("setup-guide-enable-title").innerHTML = _("Guide")+":";
 		document.getElementById("setup-guide-enable-label").innerHTML = _("Toggle guide");
 		document.getElementById("setup-guide-enable").title = _("Toggle guide");
-		document.getElementById("setup-downloadImage").innerHTML = "<b>"+_("Download whiteboard image")+"</b>";
-		document.getElementById("setup-downloadLayers").innerHTML = "<b>"+_("Download layers")+"</b>";
-		document.getElementById("setup-downloadLayers-interval-title").innerHTML = _("Interval (in msecs)")+": ";
+		document.getElementById("setup-downloadLayers-title").innerHTML = _("Download layers")+": ";
+		document.getElementById("setup-downloadLayers-interval-title").innerHTML = _("Animation interval (in msecs)")+": ";
+		document.getElementById("setup-downloadLayers-columns-title").innerHTML = _("Layers per row")+": ";
 		document.getElementById("setup-execute-time-title").innerHTML = _("Stop execution after")+" ";
 		document.getElementById("setup-execute-time").title = _("Number of seconds to run");
 		document.getElementById("setup-execute-time-title2").innerHTML = " "+_("seconds");
+		document.getElementById("whiteboard-tabs-download-button").innerHTML = _("Download as an image");
 	}
 
 	/**